refactor(interior): 建物ローカル座標のスカラを consts.Tile に統一する
interior の Vec/Rect が持つタイル座標のスカラ型を、overworld と同じ
consts.Tile に揃える。Vec = consts.Coord[consts.Tile]、Rect のフィールドも
consts.Tile とし、独自の座標ユニットを廃する。

overworld は既に chunk 局所座標に consts.Tile を使い、offsetX の足し算で
全域フレームへ移している。interior だけ別ユニットを持つと境界でコンストラクタや
変換が要り、コードベースの流儀から外れる。同じ consts.Tile にすることで境界の
interior.Rect/Vec 構築は素の合成になり、consts.Tile() の変換が消える。

純粋な型の統一で生成ロジックは不変。ゴールデンは GOLDIE_UPDATE 無しで一致する。

+++ internal/mapplanner/interior/house_test.go
@@ -57,7 +57,7 @@ func TestPlanHouse_玄関と廊下と水回りの小部屋を持つ(t *testing.T
 
 	area := map[roleName]int{}
 	for _, hr := range plan {
-		area[hr.Role] = hr.Room.Rect.W * hr.Room.Rect.H
+		area[hr.Role] = int(hr.Room.Rect.W * hr.Room.Rect.H)
 	}
 	for _, role := range []roleName{"genkan", "corridor", "living", "kitchen", "bedroom", "bath", "toilet"} {
 		assert.Containsf(t, area, role, "役割 %s の部屋がある", role)
